change idandroid

+++ app/src/main/java/com/project/unice/embeddedprogproject/models/Contact.java
@@ -9,7 +9,7 @@ public class Contact extends AbstractModel {
     public int id;
 
     @ModelAnnotation(columnName = "IdContactAndroid", sqlType = SqliteTypes.INTEGER)
-    public int idContactAndroid;
+    public long idContactAndroid;
 
 
     //@ModelAnnotation(columnName = "Name", sqlType = SqliteTypes.TEXT)
